map it out
